Moved user indicator related code from service-map to service-statistics
