+++ control-statistics/src/test/java/fi/nls/oskari/control/statistics/user/GetUserIndicatorsHandlerTest.java
@@ -1,10 +1,8 @@
 package fi.nls.oskari.control.statistics.user;
 
-import fi.mml.map.mapwindow.service.db.UserIndicatorService;
-import fi.mml.map.mapwindow.service.db.UserIndicatorServiceImpl;
+import org.oskari.statistics.user.UserIndicatorService;
+import org.oskari.statistics.user.UserIndicatorServiceImpl;
 import fi.nls.oskari.control.ActionParameters;
-import fi.nls.oskari.control.statistics.user.GetUserIndicatorsHandler;
-import fi.nls.oskari.control.statistics.user.SaveUserIndicatorHandler;
 import fi.nls.oskari.domain.map.indicator.UserIndicator;
 import fi.nls.test.control.JSONActionRouteTest;
 import fi.nls.test.util.ResourceHelper;
