+++ service-statistics/src/main/resources/META-INF/UserIndicator.xml
