+++ webapp-map/pom.xml
@@ -62,10 +62,6 @@
         <!-- Statistics plugins -->
         <dependency>
             <groupId>fi.nls.oskari.service</groupId>
-            <artifactId>oskari-statistics-eurostat</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>fi.nls.oskari.service</groupId>
             <artifactId>service-statistics-pxweb</artifactId>
         </dependency>
         <!-- /Statistics plugins -->
@@ -82,7 +78,6 @@
         </dependency>
         <dependency>
             <groupId>fi.nls.oskari</groupId>
-            <!-- artifactId>map-servlet</artifactId -->
             <artifactId>servlet-map</artifactId>
             <exclusions>
                 <exclusion>
